+++ MailScanner_perl_scripts/SQLBlackWhiteList.pm
@@ -41,6 +41,7 @@ use vars qw($VERSION);
 $VERSION = substr q$Revision: 1.4 $, 10;
 
 use DBI;
+use MailScanner::Config;
 my(%Whitelist, %Blacklist);
 my($wtime, $btime);
 my($refresh_time) = 15;		# Time in minutes before lists are refreshed
@@ -85,7 +86,6 @@ sub SQLBlacklist {
   return LookupList($message, \%Blacklist);
 }
 
-
 #
 # Close down the by-domain whitelist and blacklist
 #
@@ -100,15 +100,13 @@ sub EndSQLBlacklist {
 sub CreateList {
   my($type, $BlackWhite) = @_;
   my($dbh, $sth, $sql, $to_address, $from_address, $count, $filter);
-  my($db_name) = 'mailscanner';
-  my($db_host) = 'localhost';
-  my($db_user) = 'root';
-  my($db_pass) = '';
+  # Get parameters from MailScanner config file
+  my($db_dsn)  = MailScanner::Config::Value('dbdsn');
+  my($db_user) = MailScanner::Config::Value('dbusername');
+  my($db_pass) = MailScanner::Config::Value('dbpassword');
   
   # Connect to the database
-  $dbh = DBI->connect("DBI:mysql:database=$db_name;host=$db_host",
-                      $db_user, $db_pass,
-                      {PrintError => 0}); 
+  $dbh = DBI->connect($db_dsn, $db_user, $db_pass, {PrintError => 0});
 
   # Check if connection was successfull - if it isn't
   # then generate a warning and continue processing.
@@ -127,7 +125,7 @@ sub CreateList {
    $count++;
   }  
 
-    $sql = "SELECT filter, from_address FROM $type INNER JOIN user_filters ON $type.to_address = user_filters.username";
+  $sql = "SELECT filter, from_address FROM $type INNER JOIN user_filters ON $type.to_address = user_filters.username";
   $sth = $dbh->prepare($sql);
   $sth->execute;
   $sth->bind_columns(undef,\$filter,\$from_address);